feat: server.js support SPA
