+++ package.json
@@ -4,12 +4,11 @@
   "private": true,
   "main": "server.js",
   "scripts": {
-    "serve": "vue-cli-service serve",
+    "serve": "node server.js",
     "build": "vue-cli-service build",
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "body-parser": "^1.19.0",
     "express": "^4.17.1",
     "mime": "^2.5.2"
   },
