+++ src/main.js
@@ -1,11 +1,12 @@
 import Vue from "vue";
 import App from "./App.vue";
 import store from "./store.js";
+import router from "./router";
 
 Vue.config.productionTip = false;
 
-
 new Vue({
   render: (h) => h(App),
-  store: store
+  store,
+  router,
 }).$mount("#app");
